Move createSupergraphSDLFetcher to /core package
