+++ .changeset/tough-rice-own.md
@@ -1,0 +1,6 @@
+---
+'@graphql-hive/apollo': patch
+'@graphql-hive/core': patch
+---
+
+Move createSupergraphSDLFetcher to @graphql-hive/core package
