+++ packages/libraries/apollo/src/index.ts
@@ -2,85 +2,22 @@ import { GraphQLError, type DocumentNode } from 'graphql';
 import type { ApolloServerPlugin, HTTPGraphQLRequest } from '@apollo/server';
 import {
   autoDisposeSymbol,
-  createHash,
   createHive as createHiveClient,
+  createSupergraphSDLFetcher,
   HiveClient,
   HivePluginOptions,
-  http,
   isHiveClient,
-  joinUrl,
-  Logger,
+  SupergraphSDLFetcherOptions,
 } from '@graphql-hive/core';
 import { version } from './version.js';
 
-export { atLeastOnceSampler, createSchemaFetcher, createServicesFetcher } from '@graphql-hive/core';
-
-export interface SupergraphSDLFetcherOptions {
-  endpoint: string;
-  key: string;
-  logger?: Logger;
-}
-
-export function createSupergraphSDLFetcher(options: SupergraphSDLFetcherOptions) {
-  let cacheETag: string | null = null;
-  let cached: {
-    id: string;
-    supergraphSdl: string;
-  } | null = null;
-  const endpoint = options.endpoint.endsWith('/supergraph')
-    ? options.endpoint
-    : joinUrl(options.endpoint, 'supergraph');
-
-  return function supergraphSDLFetcher(): Promise<{ id: string; supergraphSdl: string }> {
-    const headers: {
-      [key: string]: string;
-    } = {
-      'X-Hive-CDN-Key': options.key,
-      'User-Agent': `hive-client/${version}`,
-    };
-
-    if (cacheETag) {
-      headers['If-None-Match'] = cacheETag;
-    }
-
-    return http
-      .get(endpoint, {
-        headers,
-        isRequestOk: response => response.status === 304 || response.ok,
-        retry: {
-          retries: 10,
-          maxTimeout: 200,
-          minTimeout: 1,
-        },
-        logger: options.logger,
-      })
-      .then(async response => {
-        if (response.ok) {
-          const supergraphSdl = await response.text();
-          const result = {
-            id: await createHash('SHA-256').update(supergraphSdl).digest('base64'),
-            supergraphSdl,
-          };
-
-          const etag = response.headers.get('etag');
-          if (etag) {
-            cached = result;
-            cacheETag = etag;
-          }
-
-          return result;
-        }
-
-        if (response.status === 304 && cached !== null) {
-          return cached;
-        }
-
-        throw new Error(
-          `Failed to GET ${endpoint}, received: ${response.status} ${response.statusText ?? 'Internal Server Error'}`,
-        );
-      });
-  };
-}
+export {
+  atLeastOnceSampler,
+  createSchemaFetcher,
+  createServicesFetcher,
+  createSupergraphSDLFetcher,
+} from '@graphql-hive/core';
+export type { SupergraphSDLFetcherOptions } from '@graphql-hive/core';
 
 export function createSupergraphManager(
   options: { pollIntervalInMs?: number } & SupergraphSDLFetcherOptions,
