+++ packages/libraries/core/src/version.ts
@@ -1,1 +1,1 @@
-export const version = '0.5.0';
+export const version = '0.7.0';
